+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Experimental
 ### Added
 - Native player for video-files
+- It is now possible to hide posts by hashtags! Also, the underlying algorithm
+  allows to add other types of hiding criteria in the future.
 
 ### Changed
 - Take Save out of "more" menu